adding and removing workout

+++ app/src/main/java/pl/edu/am_projekt/InputData.kt
@@ -1,6 +1,6 @@
 package pl.edu.am_projekt
 
 data class InputData(
-    var first: Int,
+    var first: Double,
     var second: Int,
 )
